refactor: 디테일 마커 리팩토링 코드 정리
